+++ src/modules/name-cards/name-cards.docs.ts
@@ -1,5 +1,5 @@
 import { applyDecorators } from '@nestjs/common';
-import { ApiOperation, ApiQuery } from '@nestjs/swagger';
+import { ApiOperation, ApiParam, ApiQuery } from '@nestjs/swagger';
 import { SwaggerMethodDoc } from '../../swagger/swagger-method-doc-type';
 import { NameCardController } from './name-cards.controller';
 
@@ -14,6 +14,18 @@ export const ApiDocs: SwaggerMethodDoc<NameCardController> = {
       }),
       ApiQuery({
         name: 'userId',
+        required: true,
+      }),
+    );
+  },
+  getNamecardByUniqueCode(summary: string) {
+    return applyDecorators(
+      ApiOperation({
+        summary,
+      }),
+      ApiParam({
+        name: 'NameCardUniqueCode',
+        description: '명함 당 부여되는 고유 코드',
         required: true,
       }),
     );
